Implement Delete and Authenticate CredHub calls
Rewrite some tests as CredHub-integrating contract tests.

A contract test allows us to get quick feedback on whether we're
compatible with new CredHub versions of e.g. CLI library or server.
Also, the test can be 'pointed' at other implementations, to allow
higher-level tests to run with a fake credential store without talking
to CredHub.

This approach also means we don't have to pretend to be the CredHub
server.

[#151586751]

Signed-off-by: Andrew Bruce <abruce@pivotal.io>