Retire own implementation of bosh director client
With this changes, we no longer rely on our own implementation of a client to
communicate with the bosh director through the Bosh API. Instead, we are now
using the bosh cli as a library and leveraging the methods implemented by the
bosh team.

On the flip side, skipping ssl verification is no longer permitted.

Signed-off-by: Andrea Nodari <anodari@pivotal.io>