Fix flake on Parallel System test
We can guarantee the order in which the service instances will start
upgrading, but we can't guarantee the order in which they will complete.

As gbytes.Say keeps track of the last byte checked, by hard-coding the
complete ordering we introduce a flake: when the second service instance
completes after the third, the test fails -- even when both complete
successfully.

[#156845448]