Don't perform any transformations on the generated manifest during updates
We accidentally introduced a new behaviour by replacing
((odb_secret:key)) with ((/odb/id/id/key)) during the check for pending
changes on updates.

This commit rolls back to the old behaviour: if the manifest that bosh
knows about is different than the manifest generated by the adapter,
update will fail. To ensure that the manifest looks the same, adapter
authors are expected to extract the ODB generated CredHub name from the
previous manifest (that's already provided to `generate-manifest`)

[#158766590]

Co-authored-by: Derik Evangelista <devangelista@pivotal.io>