ODB should not call BulkSet when bosh credhub is not configured
on `main`, we construct a deployer passing a reference to a
`credhub.Store` that is either:

1. `nil` when credhub is not configured
2. A pointer to a `credhub.Store` when credhub is configured

The deployer, however, receives an interface `BulkSetter` in its
constructor arguments. The deployer will later check whether the
BulkSetter is nil to decide whether to call it or not.

The problem lies that we cannot reliable check if an interface is nil
without using reflection. This commit implements exactly that.

A more elegant solution would be to create a Noop bulksetter and pass
that to the deployer when we the `resolve_manifest_at_bind` is false.
That would require quite a refactoring (as we would not want to call the
manifestGenerator when the BulkSetter is a Noop).

[finishes #159266330]

Co-authored-by: Jonathan Tsang <jtsang@pivotal.io>