Fix listing the configs by returning only the latest configs
Director `ListConfigs` function accepts `limit` as the 1st parameter.
When `limit` is `1`, it will return the latest config for each
type/name config, otherwise, it will return all config ids for each
type/name config.

This commit fixes the `ListConfigs` invokations:
* At `GetConfigs` we only need the latest config for each type/name,
as no other part of the code needs to deal with the history of configs.
* At `DeleteConfig` we only need the list of each type/name config,
as we are using the director `DeleteConfig` which will delete all
config ids for a particular type/name.