Rename Builder to Configurator
The instanceiterator had an object called a Builder which confused
people new to the codebase as they thought it followed the commonly
known [Builder Pattern](https://en.wikipedia.org/wiki/Builder_pattern),
even though it did not, and did not build anything at all. Hence it has
been renamed to a Configurator because it configures things, and there
is no design pattern with the same name (the Component Configurator
pattern has a sufficiently different name).

[#168833866]