Refactor - context should always be first argument
While inspecting code it was seen that the context was not always the
first argument in some functions. According to the doc for context:

"Do not store Contexts inside a struct type; instead, pass a Context
explicitly to each function that needs it. The Context should be the
first parameter, typically named ctx"

[#169177292]