chore: get NATS cert via properties script
There is a pre-existing script (not public) that extract values relating
to test environents and makes them available as a `vars.yaml` file which
is used in the tests. This is a better mechanism for getting the NATS
client certificate as it is consistent with how other certs are fetched,
and it works for different deployment names.

[#180262739]("https://www.pivotaltracker.com/story/show/180262739")