enable allowpublic feature flag in lifecycle system tests
This uaa client definition flag was enabled in #189, but this
configuration cannot actually be deployed without some release-level
templating changes.

Backfilling a system_test to drive out the templating changes. The bosh
templates currently enforce an "allowlist" of valid properties to
`uaa.client_definition` and that list needs to be updated to allow
`allowpublic`.

This system_test should pass once a change is made to this allowlist in
[pivotal-cf/on-demand-service-broker-release](https://github.com/pivotal-cf/on-demand-service-broker-release).

[#184630501](https://www.pivotaltracker.com/story/show/184630501)

Authored-by: Andrew Garner <garnera@vmware.com>