system_tests: ensure a stable broker password
Previously, the broker password was set as the first 6 bytes of a uuid.

If this 6-byte password was an valid integer sequence rendered into
yaml, the string password could differ from the actual value used by the
bosh deployment.

For example, a random uuid, `0112115F-D7FA-4617-AE7A-D62D42FBA869` could
lead to the string password "011211", but the yaml would use the value
4745 (i.e. the decimal value of the octal 11211).

Now the broker password is set to a 36 byte string representation of the
underlying uuid which should never be interpreted as an integer.