adjust logging when persisting manifests
Intentionally not testing the failure cases for compressing manifests
because it would require plumbing in significant fake / mocks to
validate and we do not plan to fail on error anyway.

Happy path is validated through unit tests and integration tests.  If
compression fails, at worse a manifest does not get persisted and
if the logging is wrong troubleshooting why may be difficult.

[#186826718](https://www.pivotaltracker.com/story/show/186826718)

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>