Uses config to toggle between Persister and DisabledPersister
Based on the config option, if the feature is enabled, we will use the
real Persister to create the directory and manifests. Otherwise, use the
DisabledPersister which will do nothing.

For the cleaner, we always use the Persister - this means we can always
clean up old manifests when an instance is deleted, even if the feature
is disabled after being previously enabled

[#186826718](https://www.pivotaltracker.com/story/show/186826718)

Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>
Co-authored-by: Andrew Garner <garnera@vmware.com>