Ensure cf push of the example-service-instances-api app is checked
This failed because the app manifest was hardcoded with GOVERSION=1.21
which was no longer supported by the cloudfoundry environment.
Unfortunately the test proceeded anyway and failed later in a way that
was harder to troubleshoot.

Now the feature flags tests asserts that cf push actually succeeds - or
at least cf push exits non-zero and flags a better error message
up-front.